Refactor CLI into multiple files
