+++ setup.cfg
@@ -56,4 +56,4 @@ where = src
 
 [options.entry_points]
 console_scripts =
-    telliot = telliot_core.cli:main
+    telliot = telliot_core.cli.main:main
