--- src/telliot_core/__main__.py
@@ -1,4 +1,0 @@
-"""Runs a CLI to to interact with all `telliot_core`'s subpackages."""
-from telliot_core.cli import main
-
-main()
